requirements.txt
flask
gunicorn
pillow
numpy
tensorflow
